Update
files created: about.html and about.css

finished desktop on contact.html and scales fine down to mobile view.
Contact need small adjustments when it comes to implementing JS with error messages, and "for" attributes on input labels.

Added images in a new folder inside images called background, this is going to be used for background images on about and contact page.