Update 1.0.0
files created: script.js, blog-specific.js

Some adjustments to the css because of the fetch, i changed the container name of the new posts and the popular posts because i wanted to target them separately, but i now thaught that i can change them to the same again to make the css neater and fetch them through an id, this will be implemented in the next update of the page.

removed the boxes that was made for when styling the page, im going to delete the images that i dont need from the folder structure later.